+++ Changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+x.x.x (???)
+++++++++++++++++++++++++
+
+New features
+------------
+* Import of ECC curves in compressed form.
+
 3.4.12 (5 February 2018)
 ++++++++++++++++++++++++